Reorganize applications in different files
